% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amycltp.R
\name{amycltp}
\alias{amycltp}
\title{amycltp}
\usage{
amycltp(n, iter, lambda = 10, ...)
}
\arguments{
\item{n}{sample size}

\item{iter}{number of resamplings}

\item{lambda}{poisson distributuion parameter}

\item{...}{}
}
\value{
plot
}
\description{
creates a histogram of central tendency for a poisson distribution.
}
