% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboot2.R
\name{myboot2}
\alias{myboot2}
\title{myboot2}
\usage{
myboot2(iter = 10000, x, fun = "mean", alpha = 0.05, cx = 1.5, ...)
}
\arguments{
\item{iter}{number of itterations}

\item{x}{a random sample from the data points.}

\item{fun}{the statistic you want to test.}

\item{alpha}{the confidence level you want to test (alpha = 0.05 is a 95% ci)}

\item{...}{}
}
\value{

}
\description{
creates a histogram with a bootstrap interval.
}
\examples{
myboot2(iter=10000,sam, fun="mean",alpha=0.2, xlab="mean")

}
