% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myci.R
\name{myci}
\alias{myci}
\title{my ci}
\usage{
myci(x)
}
\arguments{
\item{x}{a vector containing data from a sample.}
}
\value{
returns a vector with two numbers, the lower and upper limits of the confidence interval.
}
\description{
Finds the 95% confidence interval of mu for a single sample.
}
\examples{
myci(x)

}
