% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymlnorm.R
\name{mymlnorm}
\alias{mymlnorm}
\title{mymlnorm}
\usage{
mymlnorm(x, mu, sig, ...)
}
\arguments{
\item{x}{a vector of data points.}

\item{...}{}
}
\value{

}
\description{
creates plot of maximum likelihood from a set of numbers
}
\examples{
mymlnorm(x=y,mu=seq(5,10,length=1000),sig=seq(0.1,4,length=1000),lwd=2,labcex=1)

}
