% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mypvalue.R
\name{mypvalue}
\alias{mypvalue}
\title{mypvalue}
\usage{
mypvalue(t0, xmax = 4, n = 20, alpha = 0.05)
}
\arguments{
\item{t0}{the value of t_calc.}

\item{n}{the number of datapoints in the sample}

\item{alpha}{the confidence interval (alpha = 0.05 is a 95% ci)}
}
\value{
returs a plot with the p value and
}
\description{
This function calculates the p-value given t_calc and produces a graph with the rejection region
}
